add watch component
